consensus, cmd, eth, core: refactor systemcontract and fix test
