--- consensus/dpos/systemcontract.go
@@ -1,197 +1,0 @@
-package dpos
-
-import (
-	"math/big"
-	"strings"
-
-	"github.com/themis-network/go-themis/accounts/abi"
-	"github.com/themis-network/go-themis/common"
-	"github.com/themis-network/go-themis/core/types"
-)
-
-const (
-	// Deployed code of system contract
-	mainSystemContractCode = "0x6080604052600436106101ac576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806304f49a3a146101b15780630e14a376146102225780632083c5f91461026b57806321f8a721146102c25780632c62ff2d146103335780632e28d0841461037c57806335f7ff631461040b57806336cb6592146104b55780633b7bfda0146105305780633e49bed0146105795780634a3270d7146105cc5780635948f733146106765780635beff993146106fc5780635ec2814514610745578063616b59f6146107885780636e899550146107d157806371bda38c14610860578063789c617c1461088d5780637ae1cfca146108d25780638c1600951461091b578063986e791a14610964578063a79413e114610a0e578063a855d4ce14610a65578063abfdcced14610aaa578063bd02d0f514610aff578063c031a18014610b44578063ca446dd914610bee578063dc97d96214610c57578063e2a4853a14610c9c578063e2b202bf14610cef578063e2e6385f14610d38578063e80a7eb914610d85578063ec672cf614610df1578063f6bb3cc414610e7d575b600080fd5b3480156101bd57600080fd5b506101e06004803603810190808035600019169060200190929190505050610ec6565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561022e57600080fd5b506102516004803603810190808035600019169060200190929190505050610ef9565b604051808215151515815260200191505060405180910390f35b34801561027757600080fd5b50610280610fed565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b3480156102ce57600080fd5b506102f16004803603810190808035600019169060200190929190505050611064565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561033f57600080fd5b5061036260048036038101908080356000191690602001909291905050506110a8565b604051808215151515815260200191505060405180910390f35b34801561038857600080fd5b506103f16004803603810190808035600019169060200190929190803590602001908201803590602001908080601f016020809104026020016040519081016040528093929190818152602001838380828437820191505050505050919291929050505061118a565b604051808215151515815260200191505060405180910390f35b34801561041757600080fd5b5061043a6004803603810190808035600019169060200190929190505050611272565b6040518080602001828103825283818151815260200191508051906020019080838360005b8381101561047a57808201518184015260208101905061045f565b50505050905090810190601f1680156104a75780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b3480156104c157600080fd5b506104ee600480360381019080803560001916906020019092919080359060200190929190505050611322565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561053c57600080fd5b5061055f600480360381019080803560001916906020019092919050505061136f565b604051808215151515815260200191505060405180910390f35b34801561058557600080fd5b506105b260048036038101908080356000191690602001909291908035906020019092919050505061138f565b604051808215151515815260200191505060405180910390f35b3480156105d857600080fd5b506105fb6004803603810190808035600019169060200190929190505050611467565b6040518080602001828103825283818151815260200191508051906020019080838360005b8381101561063b578082015181840152602081019050610620565b50505050905090810190601f1680156106685780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561068257600080fd5b506106a56004803603810190808035600019169060200190929190505050611517565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b838110156106e85780820151818401526020810190506106cd565b505050509050019250505060405180910390f35b34801561070857600080fd5b5061072b60048036038101908080356000191690602001909291905050506115c0565b604051808215151515815260200191505060405180910390f35b34801561075157600080fd5b50610786600480360381019080803573ffffffffffffffffffffffffffffffffffffffff16906020019092919050505061169e565b005b34801561079457600080fd5b506107b76004803603810190808035600019169060200190929190505050611707565b604051808215151515815260200191505060405180910390f35b3480156107dd57600080fd5b506108466004803603810190808035600019169060200190929190803590602001908201803590602001908080601f01602080910402602001604051908101604052809392919081815260200183838082843782019150505050505091929192905050506117e5565b604051808215151515815260200191505060405180910390f35b34801561086c57600080fd5b5061088b600480360381019080803590602001909291905050506118cd565b005b34801561089957600080fd5b506108bc60048036038101908080356000191690602001909291905050506119b8565b6040518082815260200191505060405180910390f35b3480156108de57600080fd5b5061090160048036038101908080356000191690602001909291905050506119d0565b604051808215151515815260200191505060405180910390f35b34801561092757600080fd5b5061094a6004803603810190808035600019169060200190929190505050611a02565b604051808215151515815260200191505060405180910390f35b34801561097057600080fd5b506109936004803603810190808035600019169060200190929190505050611ad8565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156109d35780820151818401526020810190506109b8565b50505050905090810190601f168015610a005780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b348015610a1a57600080fd5b50610a23611b95565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b348015610a7157600080fd5b50610a946004803603810190808035600019169060200190929190505050611c0c565b6040518082815260200191505060405180910390f35b348015610ab657600080fd5b50610ae56004803603810190808035600019169060200190929190803515159060200190929190505050611c24565b604051808215151515815260200191505060405180910390f35b348015610b0b57600080fd5b50610b2e6004803603810190808035600019169060200190929190505050611d0f565b6040518082815260200191505060405180910390f35b348015610b5057600080fd5b50610b736004803603810190808035600019169060200190929190505050611d34565b6040518080602001828103825283818151815260200191508051906020019080838360005b83811015610bb3578082015181840152602081019050610b98565b50505050905090810190601f168015610be05780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b348015610bfa57600080fd5b50610c3d6004803603810190808035600019169060200190929190803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050611df1565b604051808215151515815260200191505060405180910390f35b348015610c6357600080fd5b50610c866004803603810190808035600019169060200190929190505050611f02565b6040518082815260200191505060405180910390f35b348015610ca857600080fd5b50610cd5600480360381019080803560001916906020019092919080359060200190929190505050611f27565b604051808215151515815260200191505060405180910390f35b348015610cfb57600080fd5b50610d1e6004803603810190808035600019169060200190929190505050611fff565b604051808215151515815260200191505060405180910390f35b348015610d4457600080fd5b50610d83600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190803590602001909291905050506120d5565b005b348015610d9157600080fd5b50610d9a6121dd565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b83811015610ddd578082015181840152602081019050610dc2565b505050509050019250505060405180910390f35b348015610dfd57600080fd5b50610e6360048036038101908080356000191690602001909291908035906020019082018035906020019080806020026020016040519081016040528093929190818152602001838360200280828437820191505050505050919291929050505061226b565b604051808215151515815260200191505060405180910390f35b348015610e8957600080fd5b50610eac6004803603810190808035600019169060200190929190505050612353565b604051808215151515815260200191505060405180910390f35b60006020528060005260406000206000915054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515610fa757600080fd5b600080836000191660001916815260200190815260200160002060006101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905560019050919050565b600080600060405180807f73797374656d2e766f746553797374656d436f6e747261637400000000000000815250601901905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b6000806000836000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561115657600080fd5b60056000836000191660001916815260200190815260200160002060006101000a81549060ff021916905560019050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561123857600080fd5b816004600085600019166000191681526020019081526020016000209080519060200190611267929190612431565b506001905092915050565b60036020528060005260406000206000915090508054600181600116156101000203166002900480601f01602080910402602001604051908101604052809291908181526020018280546001816001161561010002031660029004801561131a5780601f106112ef5761010080835404028352916020019161131a565b820191906000526020600020905b8154815290600101906020018083116112fd57829003601f168201915b505050505081565b60016020528160005260406000208181548110151561133d57fe5b906000526020600020016000915091509054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60056020528060005260406000206000915054906101000a900460ff1681565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561143d57600080fd5b81600660008560001916600019168152602001908152602001600020819055506001905092915050565b60046020528060005260406000206000915090508054600181600116156101000203166002900480601f01602080910402602001604051908101604052809291908181526020018280546001816001161561010002031660029004801561150f5780601f106114e45761010080835404028352916020019161150f565b820191906000526020600020905b8154815290600101906020018083116114f257829003601f168201915b505050505081565b60606001600083600019166000191681526020019081526020016000208054806020026020016040519081016040528092919081815260200182805480156115b457602002820191906000526020600020905b8160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001906001019080831161156a575b50505050509050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561166e57600080fd5b600160008360001916600019168152602001908152602001600020600061169591906124b1565b60019050919050565b60078190806001815401808255809150509060018203906000526020600020016000909192909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156117b557600080fd5b60046000836000191660001916815260200190815260200160002060006117dc91906124d2565b60019050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561189357600080fd5b8160036000856000191660001916815260200190815260200160002090805190602001906118c292919061251a565b506001905092915050565b600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561197957600080fd5b60078181548110151561198857fe5b9060005260206000200160006101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905550565b60066020528060005260406000206000915090505481565b600060056000836000191660001916815260200190815260200160002060009054906101000a900460ff169050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515611ab057600080fd5b6006600083600019166000191681526020019081526020016000206000905560019050919050565b60606003600083600019166000191681526020019081526020016000208054600181600116156101000203166002900480601f016020809104026020016040519081016040528092919081815260200182805460018160011615610100020316600290048015611b895780601f10611b5e57610100808354040283529160200191611b89565b820191906000526020600020905b815481529060010190602001808311611b6c57829003601f168201915b50505050509050919050565b600080600060405180807f73797374656d2e72656753797374656d436f6e74726163740000000000000000815250601801905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b60026020528060005260406000206000915090505481565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515611cd257600080fd5b8160056000856000191660001916815260200190815260200160002060006101000a81548160ff0219169083151502179055506001905092915050565b6000600260008360001916600019168152602001908152602001600020549050919050565b60606004600083600019166000191681526020019081526020016000208054600181600116156101000203166002900480601f016020809104026020016040519081016040528092919081815260200182805460018160011615610100020316600290048015611de55780601f10611dba57610100808354040283529160200191611de5565b820191906000526020600020905b815481529060010190602001808311611dc857829003601f168201915b50505050509050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515611e9f57600080fd5b81600080856000191660001916815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506001905092915050565b6000600660008360001916600019168152602001908152602001600020549050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515611fd557600080fd5b81600260008560001916600019168152602001908152602001600020819055506001905092915050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156120ad57600080fd5b6002600083600019166000191681526020019081526020016000206000905560019050919050565b600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561218157600080fd5b8160078281548110151561219157fe5b9060005260206000200160006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505050565b6060600780548060200260200160405190810160405280929190818152602001828054801561226157602002820191906000526020600020905b8160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019060010190808311612217575b5050505050905090565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561231957600080fd5b81600160008560001916600019168152602001908152602001600020908051906020019061234892919061259a565b506001905092915050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561240157600080fd5b60036000836000191660001916815260200190815260200160002060006124289190612624565b60019050919050565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061247257805160ff19168380011785556124a0565b828001600101855582156124a0579182015b8281111561249f578251825591602001919060010190612484565b5b5090506124ad919061266c565b5090565b50805460008255906000526020600020908101906124cf919061266c565b50565b50805460018160011615610100020316600290046000825580601f106124f85750612517565b601f016020900490600052602060002090810190612516919061266c565b5b50565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061255b57805160ff1916838001178555612589565b82800160010185558215612589579182015b8281111561258857825182559160200191906001019061256d565b5b509050612596919061266c565b5090565b828054828255906000526020600020908101928215612613579160200282015b828111156126125782518260006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550916020019190600101906125ba565b5b5090506126209190612691565b5090565b50805460018160011615610100020316600290046000825580601f1061264a5750612669565b601f016020900490600052602060002090810190612668919061266c565b5b50565b61268e91905b8082111561268a576000816000905550600101612672565b5090565b90565b6126d191905b808211156126cd57600081816101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905550600101612697565b5090565b905600a165627a7a723058205155b67182c5b6c4f5f500701f717511a49cc2213094fe58e86b414bccde6a010029"
-	regSystemContractCode  = "0x60806040526004361061008e576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806304be495a146100935780631af55073146100c2578063421fca721461017d578063a30f24c414610200578063aab147b21461022f578063d385c02f14610286578063e53966b7146102b5578063e80a7eb91461032b575b600080fd5b34801561009f57600080fd5b506100a8610397565b604051808215151515815260200191505060405180910390f35b3480156100ce57600080fd5b506100d761093d565b604051808060200180602001848152602001838103835286818151815260200191508051906020019060200280838360005b83811015610124578082015181840152602081019050610109565b50505050905001838103825285818151815260200191508051906020019060200280838360005b8381101561016657808201518184015260208101905061014b565b505050509050019550505050505060405180910390f35b34801561018957600080fd5b506101e6600480360381019080803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390505050610cc0565b604051808215151515815260200191505060405180910390f35b34801561020c57600080fd5b50610215611323565b604051808215151515815260200191505060405180910390f35b34801561023b57600080fd5b50610270600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050611774565b6040518082815260200191505060405180910390f35b34801561029257600080fd5b5061029b6118c9565b604051808215151515815260200191505060405180910390f35b61031160048036038101908080359060200190820180359060200191909192939192939080359060200190820180359060200191909192939192939080359060200190820180359060200191909192939192939050505061237f565b604051808215151515815260200191505060405180910390f35b34801561033757600080fd5b5061034061360c565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b83811015610383578082015181840152602081019050610368565b505050509050019250505060405180910390f35b60006002600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156104ac57600080fd5b505af11580156104c0573d6000803e3d6000fd5b505050506040513d60208110156104d657600080fd5b81019080805190602001909291905050501415156104f357600080fd5b6001600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561060657600080fd5b505af115801561061a573d6000803e3d6000fd5b505050506040513d602081101561063057600080fd5b81019080805190602001909291905050501415151561064e57600080fd5b600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561075f57600080fd5b505af1158015610773573d6000803e3d6000fd5b505050506040513d602081101561078957600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156108ac57600080fd5b505af11580156108c0573d6000803e3d6000fd5b505050506040513d60208110156108d657600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fc7ccee8c463e1c633c3c5e935f474bf6574a54a4b7857c9aa57b4e40f855f259426040518082815260200191505060405180910390a26001905090565b60608060006060806000600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e80a7eb96040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401600060405180830381600087803b1580156109cd57600080fd5b505af11580156109e1573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f820116820180604052506020811015610a0b57600080fd5b810190808051640100000000811115610a2357600080fd5b82810190506020810184811115610a3957600080fd5b8151856020820283011164010000000082111715610a5657600080fd5b505092919050505092508251604051908082528060200260200182016040528015610a905781602001602082028038833980820191505090505b509150600090505b82518160ff161015610cad57600073ffffffffffffffffffffffffffffffffffffffff16838260ff16815181101515610acd57fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff16141515610c7e57600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5848360ff16815181101515610b4357fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610c1f57600080fd5b505af1158015610c33573d6000803e3d6000fd5b505050506040513d6020811015610c4957600080fd5b8101908080519060200190929190505050828260ff16815181101515610c6b57fe5b9060200190602002018181525050610ca0565b6000828260ff16815181101515610c9157fe5b90602001906020020181815250505b8080600101915050610a98565b8282600354955095509550505050909192565b60006001600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610dd557600080fd5b505af1158015610de9573d6000803e3d6000fd5b505050506040513d6020811015610dff57600080fd5b8101908080519060200190929190505050141515610e1c57600080fd5b600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e6e616d6500000000000000000000000000000000000000815250600d018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902089896040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b158015610f4f57600080fd5b505af1158015610f63573d6000803e3d6000fd5b505050506040513d6020811015610f7957600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e77656255726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902087876040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b1580156110be57600080fd5b505af11580156110d2573d6000803e3d6000fd5b505050506040513d60208110156110e857600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e70327055726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902085856040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b15801561122d57600080fd5b505af1158015611241573d6000803e3d6000fd5b505050506040513d602081101561125757600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167ff803dd90f36c9991471017caaf6c9b7c7fc074fae88b065bb25a3072149999888888888888886040518080602001806020018060200184810384528a8a8281815260200192508082843782019150508481038352888882818152602001925080828437820191505084810382528686828181526020019250808284378201915050995050505050505050505060405180910390a2600190509695505050505050565b600080600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561143757600080fd5b505af115801561144b573d6000803e3d6000fd5b505050506040513d602081101561146157600080fd5b810190808051906020019092919050505014151561147e57600080fd5b600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561159857600080fd5b505af11580156115ac573d6000803e3d6000fd5b505050506040513d60208110156115c257600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f78792e7374617475730000000000000000000000000000000000000000815250600c018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156116ee57600080fd5b505af1158015611702573d6000803e3d6000fd5b505050506040513d602081101561171857600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fc686b241606f36270db6f798cbe0c01c064676c4918449e684002f91af8d560c60405160405180910390a26001905090565b6000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58360405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561188757600080fd5b505af115801561189b573d6000803e3d6000fd5b505050506040513d60208110156118b157600080fd5b81019080805190602001909291905050509050919050565b60006001600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156119de57600080fd5b505af11580156119f2573d6000803e3d6000fd5b505050506040513d6020811015611a0857600080fd5b8101908080519060200190929190505050141515611a2557600080fd5b6001600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611b3857600080fd5b505af1158015611b4c573d6000803e3d6000fd5b505050506040513d6020811015611b6257600080fd5b8101908080519060200190929190505050141515611b7f57600080fd5b600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611c9057600080fd5b505af1158015611ca4573d6000803e3d6000fd5b505050506040513d6020811015611cba57600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015611de657600080fd5b505af1158015611dfa573d6000803e3d6000fd5b505050506040513d6020811015611e1057600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f64756365722e6f757454696d65000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020426040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015611f3b57600080fd5b505af1158015611f4f573d6000803e3d6000fd5b505050506040513d6020811015611f6557600080fd5b810190808051906020019092919050505050600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2e6385f60006121166001600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156120cd57600080fd5b505af11580156120e1573d6000803e3d6000fd5b505050506040513d60208110156120f757600080fd5b810190808051906020019092919050505061372a90919063ffffffff16565b6040518363ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200182815260200192505050600060405180830381600087803b15801561219b57600080fd5b505af11580156121af573d6000803e3d6000fd5b505050506007600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156122c657600080fd5b505af11580156122da573d6000803e3d6000fd5b505050506040513d60208110156122f057600080fd5b810190808051906020019092919050505090806001815401808255809150509060018203906000526020600020016000909192909190915055503373ffffffffffffffffffffffffffffffffffffffff167f3b19b509f96372cac17503fd8376ee2a44539ffd7e0951c9350df87e97036270426040518082815260200191505060405180910390a26001905090565b6000803390506000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561249857600080fd5b505af11580156124ac573d6000803e3d6000fd5b505050506040513d60208110156124c257600080fd5b81019080805190602001909291905050501415156124df57600080fd5b346000541415156124ef57600080fd5b600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561260957600080fd5b505af115801561261d573d6000803e3d6000fd5b505050506040513d602081101561263357600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e6e616d6500000000000000000000000000000000000000815250600d018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390208a8a6040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b15801561277857600080fd5b505af115801561278c573d6000803e3d6000fd5b505050506040513d60208110156127a257600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e77656255726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902088886040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b1580156128e757600080fd5b505af11580156128fb573d6000803e3d6000fd5b505050506040513d602081101561291157600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e70327055726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902086866040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b158015612a5657600080fd5b505af1158015612a6a573d6000803e3d6000fd5b505050506040513d6020811015612a8057600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e6465706f736974000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020346040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612bab57600080fd5b505af1158015612bbf573d6000803e3d6000fd5b505050506040513d6020811015612bd557600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612d0157600080fd5b505af1158015612d15573d6000803e3d6000fd5b505050506040513d6020811015612d2b57600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e6f757454696d65000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206004546040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612e5857600080fd5b505af1158015612e6c573d6000803e3d6000fd5b505050506040513d6020811015612e8257600080fd5b81019080805190602001909291905050505060006007805490501480612ec25750600060076000815481101515612eb557fe5b9060005260206000200154145b1561322057600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206130a96001600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e80a7eb96040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401600060405180830381600087803b15801561300957600080fd5b505af115801561301d573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f82011682018060405250602081101561304757600080fd5b81019080805164010000000081111561305f57600080fd5b8281019050602081018481111561307557600080fd5b815185602082028301116401000000008211171561309257600080fd5b50509291905050505161374390919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561310a57600080fd5b505af115801561311e573d6000803e3d6000fd5b505050506040513d602081101561313457600080fd5b810190808051906020019092919050505050600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635ec28145826040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001915050600060405180830381600087803b15801561320357600080fd5b505af1158015613217573d6000803e3d6000fd5b50505050613549565b600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2e6385f8261328d60016007600081548110151561327457fe5b906000526020600020015461372a90919063ffffffff16565b6040518363ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200182815260200192505050600060405180830381600087803b15801561331257600080fd5b505af1158015613326573d6000803e3d6000fd5b50505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020600760008154811015156133f157fe5b90600052602060002001546040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561345d57600080fd5b505af1158015613471573d6000803e3d6000fd5b505050506040513d602081101561348757600080fd5b810190808051906020019092919050505050600160078054905011156135115760076001600780549050038154811015156134be57fe5b9060005260206000200154600760008154811015156134d957fe5b906000526020600020018190555060076001600780549050038154811015156134fe57fe5b9060005260206000200160009055613532565b60006007600081548110151561352357fe5b90600052602060002001819055505b6007805480919060019003613547919061375f565b505b3373ffffffffffffffffffffffffffffffffffffffff167fc39145bd434b7cf595e248904b19659d40528dda5c35c3b3078c2889c75d283c898989898989346040518080602001806020018060200185815260200184810384528b8b82818152602001925080828437820191505084810383528989828181526020019250808284378201915050848103825287878281815260200192508082843782019150509a505050505050505050505060405180910390a260019150509695505050505050565b6060600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e80a7eb96040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401600060405180830381600087803b15801561369457600080fd5b505af11580156136a8573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156136d257600080fd5b8101908080516401000000008111156136ea57600080fd5b8281019050602081018481111561370057600080fd5b815185602082028301116401000000008211171561371d57600080fd5b5050929190505050905090565b600082821115151561373857fe5b818303905092915050565b6000818301905082811015151561375657fe5b80905092915050565b81548183558181111561378657818360005260206000209182019101613785919061378b565b5b505050565b6137ad91905b808211156137a9576000816000905550600101613791565b5090565b905600a165627a7a72305820c43b3d91e1cf43f1e433f94674b70ea552b1ecba007ae79fe911ff369961dbb90029"
-	voteSystemContractCode = "0x608060405260043610610078576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff168063545f4a711461007d5780635b41bc7d146100d05780639c46c9b3146100ff578063ae2e6f4914610156578063bed7960914610185578063dc1e30da146101eb575b600080fd5b34801561008957600080fd5b506100b66004803603810190808035906020019082018035906020019190919293919293905050506102c4565b604051808215151515815260200191505060405180910390f35b3480156100dc57600080fd5b506100e56113df565b604051808215151515815260200191505060405180910390f35b34801561010b57600080fd5b50610140600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050611ddd565b6040518082815260200191505060405180910390f35b34801561016257600080fd5b5061016b611f32565b604051808215151515815260200191505060405180910390f35b6101d1600480360381019080803573ffffffffffffffffffffffffffffffffffffffff16906020019092919080359060200190820180359060200191909192939192939050505061307a565b604051808215151515815260200191505060405180910390f35b3480156101f757600080fd5b5061022c600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050614717565b604051808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200180602001848152602001838152602001828103825285818151815260200191508051906020019060200280838360005b838110156102ad578082015181840152602081019050610292565b505050509050019550505050505060405180910390f35b6000806060600060028060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156103dd57600080fd5b505af11580156103f1573d6000803e3d6000fd5b505050506040513d602081101561040757600080fd5b810190808051906020019092919050505014151561042457600080fd5b61045c868680806020026020016040519081016040528093929190818152602001838360200280828437820191505050505050614cca565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561056d57600080fd5b505af1158015610581573d6000803e3d6000fd5b505050506040513d602081101561059757600080fd5b81019080805190602001909291905050509250600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7333360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b1580156106bb57600080fd5b505af11580156106cf573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156106f957600080fd5b81019080805164010000000081111561071157600080fd5b8281019050602081018481111561072757600080fd5b815185602082028301116401000000008211171561074457600080fd5b50509291905050509150600082511415610a8057600090505b85859050811015610a7b57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a87878481811015156107b457fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206109d186600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58c8c89818110151561089957fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561098857600080fd5b505af115801561099c573d6000803e3d6000fd5b505050506040513d60208110156109b257600080fd5b810190808051906020019092919050505061532d90919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015610a3257600080fd5b505af1158015610a46573d6000803e3d6000fd5b505050506040513d6020811015610a5c57600080fd5b810190808051906020019092919050505050808060010191505061075d565b61109f565b600090505b8151811015610d7b57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8383815181101515610ada57fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020610cd186600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58888815181101515610bac57fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610c8857600080fd5b505af1158015610c9c573d6000803e3d6000fd5b505050506040513d6020811015610cb257600080fd5b810190808051906020019092919050505061534990919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015610d3257600080fd5b505af1158015610d46573d6000803e3d6000fd5b505050506040513d6020811015610d5c57600080fd5b8101908080519060200190929190505050508080600101915050610a85565b600090505b8585905081101561109e57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8787848181101515610dd757fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020610ff486600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58c8c898181101515610ebc57fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610fab57600080fd5b505af1158015610fbf573d6000803e3d6000fd5b505050506040513d6020811015610fd557600080fd5b810190808051906020019092919050505061532d90919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561105557600080fd5b505af1158015611069573d6000803e3d6000fd5b505050506040513d602081101561107f57600080fd5b8101908080519060200190929190505050508080600101915050610d80565b5b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156111b957600080fd5b505af11580156111cd573d6000803e3d6000fd5b505050506040513d60208110156111e357600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ec672cf63360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902088886040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250602002808284378201915050945050505050602060405180830381600087803b15801561132b57600080fd5b505af115801561133f573d6000803e3d6000fd5b505050506040513d602081101561135557600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fd46c1e81ba0ed24eccce499e94541a147c4fa3f6346854d76d70f3b6d401f8db8787604051808060200182810382528484828181526020019250602002808284378201915050935050505060405180910390a26001935050505092915050565b600080606060028060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156114f657600080fd5b505af115801561150a573d6000803e3d6000fd5b505050506040513d602081101561152057600080fd5b810190808051906020019092919050505014151561153d57600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561165057600080fd5b505af1158015611664573d6000803e3d6000fd5b505050506040513d602081101561167a57600080fd5b810190808051906020019092919050505014151561169757600080fd5b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156117a857600080fd5b505af11580156117bc573d6000803e3d6000fd5b505050506040513d60208110156117d257600080fd5b81019080805190602001909291905050509150600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7333360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b1580156118f657600080fd5b505af115801561190a573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f82011682018060405250602081101561193457600080fd5b81019080805164010000000081111561194c57600080fd5b8281019050602081018481111561196257600080fd5b815185602082028301116401000000008211171561197f57600080fd5b505092919050505090506119968133846000615362565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635beff9933360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611aa757600080fd5b505af1158015611abb573d6000803e3d6000fd5b505050506040513d6020811015611ad157600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015611bfd57600080fd5b505af1158015611c11573d6000803e3d6000fd5b505050506040513d6020811015611c2757600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611d4a57600080fd5b505af1158015611d5e573d6000803e3d6000fd5b505050506040513d6020811015611d7457600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167f46f83f2216482f0d8839e93d2c2616a5fb914d68342883e58d8842b441bd54d4426040518082815260200191505060405180910390a260019250505090565b6000600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611ef057600080fd5b505af1158015611f04573d6000803e3d6000fd5b505050506040513d6020811015611f1a57600080fd5b81019080805190602001909291905050509050919050565b600080600060606003600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561204c57600080fd5b505af1158015612060573d6000803e3d6000fd5b505050506040513d602081101561207657600080fd5b810190808051906020019092919050505014151561209357600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156121a657600080fd5b505af11580156121ba573d6000803e3d6000fd5b505050506040513d60208110156121d057600080fd5b81019080805190602001909291905050501415156121ed57600080fd5b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156122fe57600080fd5b505af1158015612312573d6000803e3d6000fd5b505050506040513d602081101561232857600080fd5b81019080805190602001909291905050509250600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166321f8a7213360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561244c57600080fd5b505af1158015612460573d6000803e3d6000fd5b505050506040513d602081101561247657600080fd5b81019080805190602001909291905050509150600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614151561292b57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206126d486600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58860405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561268b57600080fd5b505af115801561269f573d6000803e3d6000fd5b505050506040513d60208110156126b557600080fd5b810190808051906020019092919050505061534990919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561273557600080fd5b505af1158015612749573d6000803e3d6000fd5b505050506040513d602081101561275f57600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b15801561288257600080fd5b505af1158015612896573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156128c057600080fd5b8101908080516401000000008111156128d857600080fd5b828101905060208101848111156128ee57600080fd5b815185602082028301116401000000008211171561290b57600080fd5b505092919050505090506129228183856000615362565b60019350613074565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7333360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b158015612a3c57600080fd5b505af1158015612a50573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f820116820180604052506020811015612a7a57600080fd5b810190808051640100000000811115612a9257600080fd5b82810190506020810184811115612aa857600080fd5b8151856020820283011164010000000082111715612ac557600080fd5b50509291905050509050612adc8133856000615362565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612bf657600080fd5b505af1158015612c0a573d6000803e3d6000fd5b505050506040513d6020811015612c2057600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e756e766f746554696d650000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020426040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612d4b57600080fd5b505af1158015612d5f573d6000803e3d6000fd5b505050506040513d6020811015612d7557600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635beff9933360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612e9857600080fd5b505af1158015612eac573d6000803e3d6000fd5b505050506040513d6020811015612ec257600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16630e14a3763360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612fe557600080fd5b505af1158015612ff9573d6000803e3d6000fd5b505050506040513d602081101561300f57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167f2199fcd2dd4d48edeef57d851b15e6ef618dd40ae1104d8e753602502e33ef1c426040518082815260200191505060405180910390a2600193505b50505090565b6000806060600080600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561319357600080fd5b505af11580156131a7573d6000803e3d6000fd5b505050506040513d60208110156131bd57600080fd5b81019080805190602001909291905050501415156131da57600080fd5b60015434101515156131eb57600080fd5b600073ffffffffffffffffffffffffffffffffffffffff168773ffffffffffffffffffffffffffffffffffffffff1614151561365a5760028060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58960405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561333357600080fd5b505af1158015613347573d6000803e3d6000fd5b505050506040513d602081101561335d57600080fd5b810190808051906020019092919050505014151561337a57600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58960405180807f70726f78792e7374617475730000000000000000000000000000000000000000815250600c018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561348d57600080fd5b505af11580156134a1573d6000803e3d6000fd5b505050506040513d60208110156134b757600080fd5b81019080805190602001909291905050501415156134d457600080fd5b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ca446dd93360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020896040518363ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018083600019166000191681526020018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200192505050602060405180830381600087803b15801561361957600080fd5b505af115801561362d573d6000803e3d6000fd5b505050506040513d602081101561364357600080fd5b810190808051906020019092919050505050613805565b613692868680806020026020016040519081016040528093929190818152602001838360200280828437820191505050505050614cca565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ec672cf63360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902088886040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250602002808284378201915050945050505050602060405180830381600087803b1580156137c857600080fd5b505af11580156137dc573d6000803e3d6000fd5b505050506040513d60208110156137f257600080fd5b8101908080519060200190929190505050505b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060036040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561391f57600080fd5b505af1158015613933573d6000803e3d6000fd5b505050506040513d602081101561394957600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e7374616b6564000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020346040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613a7457600080fd5b505af1158015613a88573d6000803e3d6000fd5b505050506040513d6020811015613a9e57600080fd5b810190808051906020019092919050505050613ab9346157ee565b9250600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020856040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613bd457600080fd5b505af1158015613be8573d6000803e3d6000fd5b505050506040513d6020811015613bfe57600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613d2a57600080fd5b505af1158015613d3e573d6000803e3d6000fd5b505050506040513d6020811015613d5457600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e756e766f746554696d650000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000546040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613e8157600080fd5b505af1158015613e95573d6000803e3d6000fd5b505050506040513d6020811015613eab57600080fd5b810190808051906020019092919050505050600073ffffffffffffffffffffffffffffffffffffffff168773ffffffffffffffffffffffffffffffffffffffff1614151561435b57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8860405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902061410886600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58d60405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156140bf57600080fd5b505af11580156140d3573d6000803e3d6000fd5b505050506040513d60208110156140e957600080fd5b810190808051906020019092919050505061532d90919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561416957600080fd5b505af115801561417d573d6000803e3d6000fd5b505050506040513d602081101561419357600080fd5b810190808051906020019092919050505050600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338860405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b1580156142b657600080fd5b505af11580156142ca573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156142f457600080fd5b81019080805164010000000081111561430c57600080fd5b8281019050602081018481111561432257600080fd5b815185602082028301116401000000008211171561433f57600080fd5b505092919050505091506143568288856001615362565b61467f565b600090505b8585905081101561467e57600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a87878481811015156143b757fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206145d486600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58c8c89818110151561449c57fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561458b57600080fd5b505af115801561459f573d6000803e3d6000fd5b505050506040513d60208110156145b557600080fd5b810190808051906020019092919050505061532d90919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561463557600080fd5b505af1158015614649573d6000803e3d6000fd5b505050506040513d602081101561465f57600080fd5b8101908080519060200190929190505050508080600101915050614360565b5b8673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167f5fd2438cc54f76807d5a8071c168927902854cef6c0c22176ded5bbe7174a9fb88883460405180806020018381526020018281038252858582818152602001925060200280828437820191505094505050505060405180910390a3600193505050509392505050565b6000606060008060006060600080600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166321f8a7218a60405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561483657600080fd5b505af115801561484a573d6000803e3d6000fd5b505050506040513d602081101561486057600080fd5b81019080805190602001909291905050509350600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338a60405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b15801561498457600080fd5b505af1158015614998573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156149c257600080fd5b8101908080516401000000008111156149da57600080fd5b828101905060208101848111156149f057600080fd5b8151856020820283011164010000000082111715614a0d57600080fd5b50509291905050509250600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58a60405180807f766f74652e7374616b6564000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614b2857600080fd5b505af1158015614b3c573d6000803e3d6000fd5b505050506040513d6020811015614b5257600080fd5b81019080805190602001909291905050509150600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58a60405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614c7657600080fd5b505af1158015614c8a573d6000803e3d6000fd5b505050506040513d6020811015614ca057600080fd5b81019080805190602001909291905050509050838383839750975097509750505050509193509193565b6000601e825111151515614cdd57600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5846000815181101515614d2c57fe5b9060200190602002015160405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614e0857600080fd5b505af1158015614e1c573d6000803e3d6000fd5b505050506040513d6020811015614e3257600080fd5b8101908080519060200190929190505050141515614e4f57600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5846000815181101515614e9e57fe5b9060200190602002015160405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614f7a57600080fd5b505af1158015614f8e573d6000803e3d6000fd5b505050506040513d6020811015614fa457600080fd5b8101908080519060200190929190505050141515614fc157600080fd5b600190505b8151811015615329576001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5848481518110151561501d57fe5b9060200190602002015160405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156150f957600080fd5b505af115801561510d573d6000803e3d6000fd5b505050506040513d602081101561512357600080fd5b810190808051906020019092919050505014151561514057600080fd5b6001600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5848481518110151561518e57fe5b9060200190602002015160405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561526a57600080fd5b505af115801561527e573d6000803e3d6000fd5b505050506040513d602081101561529457600080fd5b81019080805190602001909291905050501415156152b157600080fd5b81818151811015156152bf57fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff1682600183038151811015156152f057fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff1610151561531c57600080fd5b8080600101915050614fc6565b5050565b6000818301905082811015151561534057fe5b80905092915050565b600082821115151561535757fe5b818303905092915050565b600080600090505b85518110156157e65782156154f5576154ee84600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f589858151811015156153c957fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156154a557600080fd5b505af11580156154b9573d6000803e3d6000fd5b505050506040513d60208110156154cf57600080fd5b810190808051906020019092919050505061532d90919063ffffffff16565b915061566d565b61566a84600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5898581518110151561554557fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561562157600080fd5b505af1158015615635573d6000803e3d6000fd5b505050506040513d602081101561564b57600080fd5b810190808051906020019092919050505061534990919063ffffffff16565b91505b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a87838151811015156156b957fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020846040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561579d57600080fd5b505af11580156157b1573d6000803e3d6000fd5b505050506040513d60208110156157c757600080fd5b810190808051906020019092919050505050808060010191505061536a565b505050505050565b600061580460018361580b90919063ffffffff16565b9050919050565b60008083141561581e576000905061583d565b818302905081838281151561582f57fe5b0414151561583957fe5b8090505b929150505600a165627a7a72305820c8a9473eee0c8a6bcb5d1b21a4413c1a52f4cb4dc913c5bb4b05484800e3f44b0029"
-
-	// System contract abi generated by abigen
-	mainSystemContractABI = "[{\"constant\":false,\"inputs\":[],\"name\":\"getVoteSystemContract\",\"outputs\":[{\"name\":\"\",\"type\":\"address\"}],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"constant\":false,\"inputs\":[],\"name\":\"getRegSystemContract\",\"outputs\":[{\"name\":\"\",\"type\":\"address\"}],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"function\"}]"
-	regSystemContractABI  = "[{\"constant\":true,\"inputs\":[],\"name\":\"getAllProducersInfo\",\"outputs\":[{\"name\":\"\",\"type\":\"address[]\"},{\"name\":\"\",\"type\":\"uint256[]\"},{\"name\":\"\",\"type\":\"uint256\"}],\"payable\":false,\"stateMutability\":\"view\",\"type\":\"function\"}]"
-	voteSystemContractABI = ""
-
-	// Method ID
-	getRegSystemContractMethodID  = "0xa79413e1"
-	getVoteSystemContractMethodID = "0x2083c5f9"
-	getAllProducersInfoMethodID   = "0x1af55073"
-)
-
-var (
-	mainSystemContractAddr = common.BytesToAddress([]byte{9})
-	regSystemContractAddr  = common.BytesToAddress([]byte{10})
-	voteSystemContractAddr = common.BytesToAddress([]byte{11})
-)
-
-// Upgradeable system contract for dpos
-// Producer/Voters can send tx to system contract for reg, unreg, vote and so on.
-var HardcodedContractsDpos = []HardcodedContract{
-	NewSystemContract(
-		mainSystemContractAddr,
-		mainSystemContractCode,
-		mainSystemContractABI,
-		// Set reg and vote system contract address
-		map[common.Hash]common.Hash{
-			// Set reg system contract bool; mapping key is : 0x3742e890ccd7b17f07ea3d26ddabc6a24bd77cb109451e8c225eced73747e219
-			common.HexToHash("0xddefb938278018c16c1e429eff219977102bbd1a16c0a9ca12f10ae5d425e820"): common.BytesToHash([]byte{1}),
-			// Set vote system contract bool; mapping key is : 0x46a34feee9b84783782542ca731df2e0a2b7305bf7c1dd556e25d64fd9fa0bd2
-			common.HexToHash("0xc6eca4648b4c9f4004f098d97590d949321f0016aa3c54114fe524197e35c089"): common.BytesToHash([]byte{1}),
-			// Set reg system contract address
-			common.HexToHash("0x11cde89d6b0e279583592f4b5d7e2a720237412a0a98c492b67e3246c187c00f"): common.BytesToHash([]byte{10}),
-			// Set vote system contract address
-			common.HexToHash("0x31e02a136b0147c5e92de4e9eaf91014b31370d815bccd7fef5f673bbae5f4ad"): common.BytesToHash([]byte{11}),
-		},
-	),
-	NewSystemContract(
-		regSystemContractAddr,
-		regSystemContractCode,
-		regSystemContractABI,
-		map[common.Hash]common.Hash{
-			// Set init variables
-			common.BytesToHash([]byte{0}): common.HexToHash("0x0000000000000000000000000000000000000000000000000de0b6b3a7640000"),
-			common.BytesToHash([]byte{1}): common.HexToHash("0x000000000000000000000000000000000000000000000000000000000003f480"),
-			common.BytesToHash([]byte{2}): common.HexToHash("0x000000000000000000000000000000000000000000000000000000000003f480"),
-			common.BytesToHash([]byte{3}): common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000004"),
-			common.BytesToHash([]byte{4}): common.HexToHash("0x00000000000000000000000000000000000000000000000000000007915e5b80"),
-			common.BytesToHash([]byte{5}): common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009"),
-			common.BytesToHash([]byte{6}): common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009"),
-		},
-	),
-	NewSystemContract(
-		voteSystemContractAddr,
-		voteSystemContractCode,
-		voteSystemContractABI,
-		map[common.Hash]common.Hash{
-			// Set init variables
-			common.BytesToHash([]byte{0}): common.HexToHash("0x00000000000000000000000000000000000000000000000000000007915e5b80"),
-			common.BytesToHash([]byte{1}): common.HexToHash("0x0000000000000000000000000000000000000000000000000de0b6b3a7640000"),
-			common.BytesToHash([]byte{2}): common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009"),
-		},
-	),
-}
-
-type HardcodedContract interface {
-	GetCode() string
-	GetContractAddr() common.Address
-	GetStorage() map[common.Hash]common.Hash
-}
-
-func NewSystemContract(addr common.Address, code string, abiStr string, storageSet map[common.Hash]common.Hash) *SystemContract {
-	// Ignore error
-	parsed, _ := abi.JSON(strings.NewReader(abiStr))
-	return &SystemContract{
-		Addr:       addr,
-		Code:       code,
-		StorageSet: storageSet,
-		ABI:        parsed,
-	}
-}
-
-// Main dpos upgradeable system contract entry.
-// Producer can send tx to this contract for proposing upgrade system
-// contract, the storage data will leave on this contract.
-type SystemContract struct {
-	Addr       common.Address
-	Code       string
-	StorageSet map[common.Hash]common.Hash
-	ABI        abi.ABI
-}
-
-func (m *SystemContract) GetCode() string {
-	return m.Code
-}
-
-func (m *SystemContract) GetContractAddr() common.Address {
-	return m.Addr
-}
-
-func (m *SystemContract) GetStorage() map[common.Hash]common.Hash {
-	return m.StorageSet
-}
-
-func (m *SystemContract) GetRegSystemContract(data []byte) common.Address {
-	var res = new(common.Address)
-	m.ABI.Unpack(res, "getCurrentSystemContractAddress", data)
-	return *res
-}
-
-// Caller for system contract which can get current reg system contract and get all producers.
-type SystemContractCaller struct {
-	mainSystemContractABI abi.ABI
-	regSystemContractABI  abi.ABI
-}
-
-func NewSystemContractCaller(mainABIStr, regABIStr string) *SystemContractCaller {
-	// Ignore error
-	mainABI, _ := abi.JSON(strings.NewReader(mainABIStr))
-	regABI, _ := abi.JSON(strings.NewReader(regABIStr))
-	return &SystemContractCaller{
-		mainABI,
-		regABI,
-	}
-}
-
-func (s *SystemContractCaller) GetRegSystemContractCall(header *types.Header) SystemCall {
-	copyAddr := mainSystemContractAddr
-	return NewCallMsg(&copyAddr, common.Hex2Bytes(getRegSystemContractMethodID), header.Number.Uint64())
-}
-
-func (s *SystemContractCaller) GetRegSystemContractAddress(data []byte) common.Address {
-	var res = new(common.Address)
-	s.mainSystemContractABI.Unpack(res, "getRegSystemContract", data)
-	return *res
-}
-
-func (s *SystemContractCaller) GetAllProducersInfoCall(header *types.Header, contract *common.Address) SystemCall {
-	return NewCallMsg(contract, common.Hex2Bytes(getAllProducersInfoMethodID), header.Number.Uint64())
-}
-
-func (s *SystemContractCaller) GetAllProducersInfo(data []byte) ([]common.Address, []*big.Int, *big.Int, error) {
-	var (
-		ret0 = new([]common.Address)
-		ret1 = new([]*big.Int)
-		ret2 = new(*big.Int)
-	)
-	out := &[]interface{}{
-		ret0,
-		ret1,
-		ret2,
-	}
-
-	err := s.regSystemContractABI.Unpack(out, "getAllProducersInfo", data)
-	return *ret0, *ret1, *ret2, err
-}
-
-func NewCallMsg(addr *common.Address, data []byte, blockNum uint64) SystemCall {
-	return SystemCall{
-		ToAddr:    addr,
-		DataField: data,
-		GasSupply: 50000000,
-		AtBlock:   blockNum,
-	}
-}
-
-type SystemCall struct {
-	ToAddr    *common.Address
-	DataField []byte
-	GasSupply uint64
-	AtBlock   uint64
-}
-
-func (s SystemCall) From() common.Address { return common.Address{} }
-func (s SystemCall) Nonce() uint64        { return 0 }
-func (s SystemCall) CheckNonce() bool     { return false }
-func (s SystemCall) To() *common.Address  { return s.ToAddr }
-func (s SystemCall) GasPrice() *big.Int   { return new(big.Int).SetUint64(0) }
-func (s SystemCall) Gas() uint64          { return s.GasSupply }
-func (s SystemCall) Value() *big.Int      { return new(big.Int).SetUint64(0) }
-func (s SystemCall) Data() []byte         { return s.DataField }
