+++ core/systemcontract.go
@@ -1,0 +1,246 @@
+package core
+
+import (
+    "math/big"
+    "strings"
+    
+    "github.com/themis-network/go-themis/accounts/abi"
+    "github.com/themis-network/go-themis/common"
+    "github.com/themis-network/go-themis/core/types"
+)
+
+const (
+    // Deployed code of system contract
+    mainSystemContractCode = "0x6080604052600436106101c2576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680630481c75f146101c757806304f49a3a146101f25780630e14a376146102635780632083c5f9146102ac57806321f8a721146103035780632c62ff2d146103745780632e28d084146103bd57806335f7ff631461044c57806336cb6592146104f65780633b7bfda0146105715780633e49bed0146105ba5780633f185b6c1461060d5780634a3270d7146106505780634b9f5c98146106fa5780635766b1a5146107295780635948f733146107765780635beff993146107fc578063616b59f6146108455780636e8995501461088e578063789c617c1461091d5780637ae1cfca146109625780638c160095146109ab578063986e791a146109f4578063a79413e114610a9e578063a855d4ce14610af5578063abfdcced14610b3a578063bd02d0f514610b8f578063c031a18014610bd4578063ca446dd914610c7e578063dc97d96214610ce7578063ddf8016614610d2c578063e2a4853a14610d6f578063e2b202bf14610dc2578063e80a7eb914610e0b578063ec672cf614610e77578063f6bb3cc414610f03575b600080fd5b3480156101d357600080fd5b506101dc610f4c565b6040518082815260200191505060405180910390f35b3480156101fe57600080fd5b506102216004803603810190808035600019169060200190929190505050610f59565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561026f57600080fd5b506102926004803603810190808035600019169060200190929190505050610f8c565b604051808215151515815260200191505060405180910390f35b3480156102b857600080fd5b506102c1611080565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561030f57600080fd5b5061033260048036038101908080356000191690602001909291905050506110f7565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561038057600080fd5b506103a3600480360381019080803560001916906020019092919050505061113b565b604051808215151515815260200191505060405180910390f35b3480156103c957600080fd5b506104326004803603810190808035600019169060200190929190803590602001908201803590602001908080601f016020809104026020016040519081016040528093929190818152602001838380828437820191505050505050919291929050505061121d565b604051808215151515815260200191505060405180910390f35b34801561045857600080fd5b5061047b6004803603810190808035600019169060200190929190505050611305565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156104bb5780820151818401526020810190506104a0565b50505050905090810190601f1680156104e85780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561050257600080fd5b5061052f6004803603810190808035600019169060200190929190803590602001909291905050506113b5565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561057d57600080fd5b506105a06004803603810190808035600019169060200190929190505050611402565b604051808215151515815260200191505060405180910390f35b3480156105c657600080fd5b506105f3600480360381019080803560001916906020019092919080359060200190929190505050611422565b604051808215151515815260200191505060405180910390f35b34801561061957600080fd5b5061064e600480360381019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506114fa565b005b34801561065c57600080fd5b5061067f6004803603810190808035600019169060200190929190505050611818565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156106bf5780820151818401526020810190506106a4565b50505050905090810190601f1680156106ec5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561070657600080fd5b506107276004803603810190808035151590602001909291905050506118c8565b005b34801561073557600080fd5b5061077460048036038101908080359060200190929190803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050611bbe565b005b34801561078257600080fd5b506107a56004803603810190808035600019169060200190929190505050611f2a565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b838110156107e85780820151818401526020810190506107cd565b505050509050019250505060405180910390f35b34801561080857600080fd5b5061082b6004803603810190808035600019169060200190929190505050611fd3565b604051808215151515815260200191505060405180910390f35b34801561085157600080fd5b5061087460048036038101908080356000191690602001909291905050506120b1565b604051808215151515815260200191505060405180910390f35b34801561089a57600080fd5b506109036004803603810190808035600019169060200190929190803590602001908201803590602001908080601f016020809104026020016040519081016040528093929190818152602001838380828437820191505050505050919291929050505061218f565b604051808215151515815260200191505060405180910390f35b34801561092957600080fd5b5061094c6004803603810190808035600019169060200190929190505050612277565b6040518082815260200191505060405180910390f35b34801561096e57600080fd5b50610991600480360381019080803560001916906020019092919050505061228f565b604051808215151515815260200191505060405180910390f35b3480156109b757600080fd5b506109da60048036038101908080356000191690602001909291905050506122c1565b604051808215151515815260200191505060405180910390f35b348015610a0057600080fd5b50610a236004803603810190808035600019169060200190929190505050612397565b6040518080602001828103825283818151815260200191508051906020019080838360005b83811015610a63578082015181840152602081019050610a48565b50505050905090810190601f168015610a905780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b348015610aaa57600080fd5b50610ab3612454565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b348015610b0157600080fd5b50610b2460048036038101908080356000191690602001909291905050506124cb565b6040518082815260200191505060405180910390f35b348015610b4657600080fd5b50610b7560048036038101908080356000191690602001909291908035151590602001909291905050506124e3565b604051808215151515815260200191505060405180910390f35b348015610b9b57600080fd5b50610bbe60048036038101908080356000191690602001909291905050506125ce565b6040518082815260200191505060405180910390f35b348015610be057600080fd5b50610c0360048036038101908080356000191690602001909291905050506125f3565b6040518080602001828103825283818151815260200191508051906020019080838360005b83811015610c43578082015181840152602081019050610c28565b50505050905090810190601f168015610c705780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b348015610c8a57600080fd5b50610ccd6004803603810190808035600019169060200190929190803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506126b0565b604051808215151515815260200191505060405180910390f35b348015610cf357600080fd5b50610d1660048036038101908080356000191690602001909291905050506127c1565b6040518082815260200191505060405180910390f35b348015610d3857600080fd5b50610d6d600480360381019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506127e6565b005b348015610d7b57600080fd5b50610da86004803603810190808035600019169060200190929190803590602001909291905050506128ec565b604051808215151515815260200191505060405180910390f35b348015610dce57600080fd5b50610df160048036038101908080356000191690602001909291905050506129c4565b604051808215151515815260200191505060405180910390f35b348015610e1757600080fd5b50610e20612a9a565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b83811015610e63578082015181840152602081019050610e48565b505050509050019250505060405180910390f35b348015610e8357600080fd5b50610ee9600480360381019080803560001916906020019092919080359060200190820180359060200190808060200260200160405190810160405280939291908181526020018383602002808284378201915050505050509192919290505050612b28565b604051808215151515815260200191505060405180910390f35b348015610f0f57600080fd5b50610f326004803603810190808035600019169060200190929190505050612c10565b604051808215151515815260200191505060405180910390f35b6000600f80549050905090565b60006020528060005260406000206000915054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561103a57600080fd5b600080836000191660001916815260200190815260200160002060006101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905560019050919050565b600080600060405180807f73797374656d2e766f746553797374656d436f6e747261637400000000000000815250601901905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b6000806000836000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156111e957600080fd5b60056000836000191660001916815260200190815260200160002060006101000a81549060ff021916905560019050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156112cb57600080fd5b8160046000856000191660001916815260200190815260200160002090805190602001906112fa9291906130ce565b506001905092915050565b60036020528060005260406000206000915090508054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156113ad5780601f10611382576101008083540402835291602001916113ad565b820191906000526020600020905b81548152906001019060200180831161139057829003601f168201915b505050505081565b6001602052816000526040600020818154811015156113d057fe5b906000526020600020016000915091509054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60056020528060005260406000206000915054906101000a900460ff1681565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156114d057600080fd5b81600660008560001916600019168152602001908152602001600020819055506001905092915050565b600080600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff1615156115a957600080fd5b6000600f805490501115156115bd57600080fd5b600260008460405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020600019166000191681526020019081526020016000205491506001600f80549050039050600082101580156116735750600f8054905082105b151561167e57600080fd5b8082141561169a5780600f81611694919061314e565b50611813565b600f818154811015156116a957fe5b9060005260206000200160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16600f838154811015156116e357fe5b9060005260206000200160006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508160026000600f8481548110151561173f57fe5b9060005260206000200160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002081905550600f805480919060019003611811919061314e565b505b505050565b60046020528060005260406000206000915090508054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156118c05780601f10611895576101008083540402835291602001916118c0565b820191906000526020600020905b8154815290600101906020018083116118a357829003601f168201915b505050505081565b60006001600260003360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020600019166000191681526020019081526020016000205414151561196c57600080fd5b600860000160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16141515156119cc57600080fd5b60001515600860050160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060009054906101000a900460ff161515141515611a2e57600080fd5b6001600860050160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060006101000a81548160ff02191690831515021790555060086006013390806001815401808255809150509060018203906000526020600020016000909192909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550508115611b6857600160086004015401600860040181905550600160036002600f8054905002811515611b2057fe5b0401905080600860040154101515611b6757611b66600860020154600860030160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16612cee565b5b5b3373ffffffffffffffffffffffffffffffffffffffff167f9186fc713232ad3af09e515527e73f725227602b71d227d1702cbb812da183fa83604051808215151515815260200191505060405180910390a25050565b606060006001600260003360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002054141515611c6457600080fd5b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1614151515611ca057600080fd5b6000841480611caf5750600184145b1515611cba57600080fd5b6008600101544260075401111515611cd157600080fd5b33600860000160006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550426008600101819055508360086002018190555082600860030160006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555060016008600401819055506008600601805480602002602001604051908101604052809291908181526020018280548015611dfd57602002820191906000526020600020905b8160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019060010190808311611db3575b50505050509150600090505b8151811015611e91576000600860050160008484815181101515611e2957fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060006101000a81548160ff0219169083151502179055508080600101915050611e09565b60086006016000611ea2919061317a565b3373ffffffffffffffffffffffffffffffffffffffff167f8a88481ed0fe4d97cd98dfb70bee8a75fc7333682788ddf1be2ba3897ccf8ede8585604051808381526020018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019250505060405180910390a250505050565b6060600160008360001916600019168152602001908152602001600020805480602002602001604051908101604052809291908181526020018280548015611fc757602002820191906000526020600020905b8160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019060010190808311611f7d575b50505050509050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561208157600080fd5b60016000836000191660001916815260200190815260200160002060006120a8919061317a565b60019050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561215f57600080fd5b6004600083600019166000191681526020019081526020016000206000612186919061319b565b60019050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561223d57600080fd5b81600360008560001916600019168152602001908152602001600020908051906020019061226c9291906131e3565b506001905092915050565b60066020528060005260406000206000915090505481565b600060056000836000191660001916815260200190815260200160002060009054906101000a900460ff169050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561236f57600080fd5b6006600083600019166000191681526020019081526020016000206000905560019050919050565b60606003600083600019166000191681526020019081526020016000208054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156124485780601f1061241d57610100808354040283529160200191612448565b820191906000526020600020905b81548152906001019060200180831161242b57829003601f168201915b50505050509050919050565b600080600060405180807f73797374656d2e72656753797374656d436f6e74726163740000000000000000815250601801905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b60026020528060005260406000206000915090505481565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561259157600080fd5b8160056000856000191660001916815260200190815260200160002060006101000a81548160ff0219169083151502179055506001905092915050565b6000600260008360001916600019168152602001908152602001600020549050919050565b60606004600083600019166000191681526020019081526020016000208054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156126a45780601f10612679576101008083540402835291602001916126a4565b820191906000526020600020905b81548152906001019060200180831161268757829003601f168201915b50505050509050919050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561275e57600080fd5b81600080856000191660001916815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506001905092915050565b6000600660008360001916600019168152602001908152602001600020549050919050565b600f80549050600260008360405180807f70726f64756365722e696e646578000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002081905550600f8190806001815401808255809150509060018203906000526020600020016000909192909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff16151561299a57600080fd5b81600260008560001916600019168152602001908152602001600020819055506001905092915050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515612a7257600080fd5b6002600083600019166000191681526020019081526020016000206000905560019050919050565b6060600f805480602002602001604051908101604052809291908181526020018280548015612b1e57602002820191906000526020600020905b8160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019060010190808311612ad4575b5050505050905090565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515612bd657600080fd5b816001600085600019166000191681526020019081526020016000209080519060200190612c05929190613263565b506001905092915050565b6000600560003360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060009054906101000a900460ff161515612cbe57600080fd5b6003600083600019166000191681526020019081526020016000206000612ce591906132ed565b60019050919050565b6000806000841415612ea75760008060405180807f73797374656d2e72656753797374656d436f6e74726163740000000000000000815250601801905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1691506000600560008460405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060006101000a81548160ff0219169083151502179055508260008060405180807f73797374656d2e72656753797374656d436f6e74726163740000000000000000815250601801905060405180910390206000191660001916815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b600184141561305d5760008060405180807f73797374656d2e766f746553797374656d436f6e747261637400000000000000815250601901905060405180910390206000191660001916815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690506000600560008360405180807f73797374656d2e61646472657373000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000191660001916815260200190815260200160002060006101000a81548160ff0219169083151502179055508260008060405180807f73797374656d2e766f746553797374656d436f6e747261637400000000000000815250601901905060405180910390206000191660001916815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b7f58ba6696cae1921e5cdc3943a320c4087848db4f4852d71fbe4f54e046414ae68484604051808381526020018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019250505060405180910390a150505050565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061310f57805160ff191683800117855561313d565b8280016001018555821561313d579182015b8281111561313c578251825591602001919060010190613121565b5b50905061314a9190613335565b5090565b815481835581811115613175578183600052602060002091820191016131749190613335565b5b505050565b50805460008255906000526020600020908101906131989190613335565b50565b50805460018160011615610100020316600290046000825580601f106131c157506131e0565b601f0160209004906000526020600020908101906131df9190613335565b5b50565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061322457805160ff1916838001178555613252565b82800160010185558215613252579182015b82811115613251578251825591602001919060010190613236565b5b50905061325f9190613335565b5090565b8280548282559060005260206000209081019282156132dc579160200282015b828111156132db5782518260006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555091602001919060010190613283565b5b5090506132e9919061335a565b5090565b50805460018160011615610100020316600290046000825580601f106133135750613332565b601f0160209004906000526020600020908101906133319190613335565b5b50565b61335791905b8082111561335357600081600090555060010161333b565b5090565b90565b61339a91905b8082111561339657600081816101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905550600101613360565b5090565b905600a165627a7a723058204252c3edbad7d504428a18f6466b61bd9a2a54550538d102e1ac9dd51646a9380029"
+    regSystemContractCode  = "0x608060405260043610610099576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806304be495a1461009e578063117df088146100cd5780631af55073146100e4578063421fca721461019f578063a30f24c414610222578063aab147b214610251578063d385c02f146102a8578063e53966b7146102d7578063e80a7eb91461034d575b600080fd5b3480156100aa57600080fd5b506100b36103b9565b604051808215151515815260200191505060405180910390f35b3480156100d957600080fd5b506100e261095f565b005b3480156100f057600080fd5b506100f96114ff565b604051808060200180602001848152602001838103835286818151815260200191508051906020019060200280838360005b8381101561014657808201518184015260208101905061012b565b50505050905001838103825285818151815260200191508051906020019060200280838360005b8381101561018857808201518184015260208101905061016d565b505050509050019550505050505060405180910390f35b3480156101ab57600080fd5b50610208600480360381019080803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390505050611873565b604051808215151515815260200191505060405180910390f35b34801561022e57600080fd5b50610237611ed6565b604051808215151515815260200191505060405180910390f35b34801561025d57600080fd5b50610292600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050612327565b6040518082815260200191505060405180910390f35b3480156102b457600080fd5b506102bd61247c565b604051808215151515815260200191505060405180910390f35b610333600480360381019080803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390803590602001908201803590602001919091929391929390505050612c54565b604051808215151515815260200191505060405180910390f35b34801561035957600080fd5b50610362613901565b6040518080602001828103825283818151815260200191508051906020019060200280838360005b838110156103a557808201518184015260208101905061038a565b505050509050019250505060405180910390f35b60006002600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156104ce57600080fd5b505af11580156104e2573d6000803e3d6000fd5b505050506040513d60208110156104f857600080fd5b810190808051906020019092919050505014151561051557600080fd5b6001600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561062857600080fd5b505af115801561063c573d6000803e3d6000fd5b505050506040513d602081101561065257600080fd5b81019080805190602001909291905050501415151561067057600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561078157600080fd5b505af1158015610795573d6000803e3d6000fd5b505050506040513d60208110156107ab57600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156108ce57600080fd5b505af11580156108e2573d6000803e3d6000fd5b505050506040513d60208110156108f857600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fc7ccee8c463e1c633c3c5e935f474bf6574a54a4b7857c9aa57b4e40f855f259426040518082815260200191505060405180910390a26001905090565b60008060003360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902092506002600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5856040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610a7a57600080fd5b505af1158015610a8e573d6000803e3d6000fd5b505050506040513d6020811015610aa457600080fd5b8101908080519060200190929190505050141515610ac157600080fd5b610c21600154600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f70726f64756365722e6f757454696d65000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610bd857600080fd5b505af1158015610bec573d6000803e3d6000fd5b505050506040513d6020811015610c0257600080fd5b8101908080519060200190929190505050613a1f90919063ffffffff16565b42111515610c2e57600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610d3f57600080fd5b505af1158015610d53573d6000803e3d6000fd5b505050506040513d6020811015610d6957600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663f6bb3cc43360405180807f70726f64756365722e6e616d6500000000000000000000000000000000000000815250600d018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610e8c57600080fd5b505af1158015610ea0573d6000803e3d6000fd5b505050506040513d6020811015610eb657600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663f6bb3cc43360405180807f70726f64756365722e77656255726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610fd957600080fd5b505af1158015610fed573d6000803e3d6000fd5b505050506040513d602081101561100357600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663f6bb3cc43360405180807f70726f64756365722e70327055726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561112657600080fd5b505af115801561113a573d6000803e3d6000fd5b505050506040513d602081101561115057600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf846040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156111fb57600080fd5b505af115801561120f573d6000803e3d6000fd5b505050506040513d602081101561122557600080fd5b8101908080519060200190929190505050503360405180807f70726f64756365722e6465706f736974000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209150600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5836040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561134b57600080fd5b505af115801561135f573d6000803e3d6000fd5b505050506040513d602081101561137557600080fd5b81019080805190602001909291905050509050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf836040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561142157600080fd5b505af1158015611435573d6000803e3d6000fd5b505050506040513d602081101561144b57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f193505050501580156114a3573d6000803e3d6000fd5b503373ffffffffffffffffffffffffffffffffffffffff167f8e9e3ccba72c5c1b74b7d52a9d2a3d43e43b0da1c0d0626df0e7c9112dd10b458242604051808381526020018281526020019250505060405180910390a2505050565b60608060006060806000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e80a7eb96040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401600060405180830381600087803b15801561158f57600080fd5b505af11580156115a3573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156115cd57600080fd5b8101908080516401000000008111156115e557600080fd5b828101905060208101848111156115fb57600080fd5b815185602082028301116401000000008211171561161857600080fd5b5050929190505050925082516040519080825280602002602001820160405280156116525781602001602082028038833980820191505090505b509150600090505b825181101561186057600073ffffffffffffffffffffffffffffffffffffffff16838281518110151561168957fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff1614151561183457600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f584838151811015156116fc57fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156117d857600080fd5b505af11580156117ec573d6000803e3d6000fd5b505050506040513d602081101561180257600080fd5b8101908080519060200190929190505050828281518110151561182157fe5b9060200190602002018181525050611853565b6000828281518110151561184457fe5b90602001906020020181815250505b808060010191505061165a565b8282600254955095509550505050909192565b60006001600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561198857600080fd5b505af115801561199c573d6000803e3d6000fd5b505050506040513d60208110156119b257600080fd5b81019080805190602001909291905050501415156119cf57600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e6e616d6500000000000000000000000000000000000000815250600d018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902089896040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b158015611b0257600080fd5b505af1158015611b16573d6000803e3d6000fd5b505050506040513d6020811015611b2c57600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e77656255726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902087876040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b158015611c7157600080fd5b505af1158015611c85573d6000803e3d6000fd5b505050506040513d6020811015611c9b57600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995503360405180807f70726f64756365722e70327055726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902085856040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b158015611de057600080fd5b505af1158015611df4573d6000803e3d6000fd5b505050506040513d6020811015611e0a57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167ff803dd90f36c9991471017caaf6c9b7c7fc074fae88b065bb25a3072149999888888888888886040518080602001806020018060200184810384528a8a8281815260200192508082843782019150508481038352888882818152602001925080828437820191505084810382528686828181526020019250808284378201915050995050505050505050505060405180910390a2600190509695505050505050565b600080600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611fea57600080fd5b505af1158015611ffe573d6000803e3d6000fd5b505050506040513d602081101561201457600080fd5b810190808051906020019092919050505014151561203157600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561214b57600080fd5b505af115801561215f573d6000803e3d6000fd5b505050506040513d602081101561217557600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f78792e7374617475730000000000000000000000000000000000000000815250600c018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156122a157600080fd5b505af11580156122b5573d6000803e3d6000fd5b505050506040513d60208110156122cb57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fc686b241606f36270db6f798cbe0c01c064676c4918449e684002f91af8d560c60405160405180910390a26001905090565b6000600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58360405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561243a57600080fd5b505af115801561244e573d6000803e3d6000fd5b505050506040513d602081101561246457600080fd5b81019080805190602001909291905050509050919050565b60006001600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561259157600080fd5b505af11580156125a5573d6000803e3d6000fd5b505050506040513d60208110156125bb57600080fd5b81019080805190602001909291905050501415156125d857600080fd5b6001600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156126eb57600080fd5b505af11580156126ff573d6000803e3d6000fd5b505050506040513d602081101561271557600080fd5b810190808051906020019092919050505014151561273257600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561284357600080fd5b505af1158015612857573d6000803e3d6000fd5b505050506040513d602081101561286d57600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561299957600080fd5b505af11580156129ad573d6000803e3d6000fd5b505050506040513d60208110156129c357600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f70726f64756365722e6f757454696d65000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020426040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612aee57600080fd5b505af1158015612b02573d6000803e3d6000fd5b505050506040513d6020811015612b1857600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16633f185b6c336040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001915050600060405180830381600087803b158015612be757600080fd5b505af1158015612bfb573d6000803e3d6000fd5b505050503373ffffffffffffffffffffffffffffffffffffffff167f3b19b509f96372cac17503fd8376ee2a44539ffd7e0951c9350df87e97036270426040518082815260200191505060405180910390a26001905090565b6000803390506000600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612d6d57600080fd5b505af1158015612d81573d6000803e3d6000fd5b505050506040513d6020811015612d9757600080fd5b8101908080519060200190929190505050141515612db457600080fd5b34600054141515612dc457600080fd5b600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015612ede57600080fd5b505af1158015612ef2573d6000803e3d6000fd5b505050506040513d6020811015612f0857600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e6e616d6500000000000000000000000000000000000000815250600d018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390208a8a6040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b15801561304d57600080fd5b505af1158015613061573d6000803e3d6000fd5b505050506040513d602081101561307757600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e77656255726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902088886040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b1580156131bc57600080fd5b505af11580156131d0573d6000803e3d6000fd5b505050506040513d60208110156131e657600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636e8995508260405180807f70726f64756365722e70327055726c0000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902086866040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250808284378201915050945050505050602060405180830381600087803b15801561332b57600080fd5b505af115801561333f573d6000803e3d6000fd5b505050506040513d602081101561335557600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e6465706f736974000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020346040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561348057600080fd5b505af1158015613494573d6000803e3d6000fd5b505050506040513d60208110156134aa57600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156135d657600080fd5b505af11580156135ea573d6000803e3d6000fd5b505050506040513d602081101561360057600080fd5b810190808051906020019092919050505050600460009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8260405180807f70726f64756365722e6f757454696d65000000000000000000000000000000008152506010018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206003546040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561372d57600080fd5b505af1158015613741573d6000803e3d6000fd5b505050506040513d602081101561375757600080fd5b810190808051906020019092919050505050600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ddf80166826040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001915050600060405180830381600087803b15801561382657600080fd5b505af115801561383a573d6000803e3d6000fd5b505050503373ffffffffffffffffffffffffffffffffffffffff167fc39145bd434b7cf595e248904b19659d40528dda5c35c3b3078c2889c75d283c898989898989346040518080602001806020018060200185815260200184810384528b8b82818152602001925080828437820191505084810383528989828181526020019250808284378201915050848103825287878281815260200192508082843782019150509a505050505050505050505060405180910390a260019150509695505050505050565b6060600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e80a7eb96040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401600060405180830381600087803b15801561398957600080fd5b505af115801561399d573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156139c757600080fd5b8101908080516401000000008111156139df57600080fd5b828101905060208101848111156139f557600080fd5b8151856020820283011164010000000082111715613a1257600080fd5b5050929190505050905090565b60008183019050828110151515613a3257fe5b809050929150505600a165627a7a7230582003719a852cfdad285fb9490b624de4850a8bcf6f320c1c0d2d6eb706c0c7feb10029"
+    voteSystemContractCode = "0x608060405260043610610078576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff168063545f4a711461007d5780635b41bc7d146100d0578063ae2e6f49146100ff578063bed796091461012e578063bed9d86114610194578063dc1e30da146101ab575b600080fd5b34801561008957600080fd5b506100b6600480360381019080803590602001908201803590602001919091929391929390505050610284565b604051808215151515815260200191505060405180910390f35b3480156100dc57600080fd5b506100e56111d8565b604051808215151515815260200191505060405180910390f35b34801561010b57600080fd5b50610114611a80565b604051808215151515815260200191505060405180910390f35b61017a600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190803590602001908201803590602001919091929391929390505050612c9d565b604051808215151515815260200191505060405180910390f35b3480156101a057600080fd5b506101a96141b8565b005b3480156101b757600080fd5b506101ec600480360381019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506148b2565b604051808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200180602001848152602001838152602001828103825285818151815260200191508051906020019060200280838360005b8381101561026d578082015181840152602081019050610252565b505050509050019550505050505060405180910390f35b60008060606000806002600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561039f57600080fd5b505af11580156103b3573d6000803e3d6000fd5b505050506040513d60208110156103c957600080fd5b81019080805190602001909291905050501415156103e657600080fd5b61041e878780806020026020016040519081016040528093929190818152602001838360200280828437820191505050505050614e65565b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561052f57600080fd5b505af1158015610543573d6000803e3d6000fd5b505050506040513d602081101561055957600080fd5b81019080805190602001909291905050509350600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7333360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b15801561067d57600080fd5b505af1158015610691573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156106bb57600080fd5b8101908080516401000000008111156106d357600080fd5b828101905060208101848111156106e957600080fd5b815185602082028301116401000000008211171561070657600080fd5b505092919050505092506000835114156109a357600090505b8686905081101561099e57868682818110151561073857fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209150600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a836108f487600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5886040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156108ab57600080fd5b505af11580156108bf573d6000803e3d6000fd5b505050506040513d60208110156108d557600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561095557600080fd5b505af1158015610969573d6000803e3d6000fd5b505050506040513d602081101561097f57600080fd5b810190808051906020019092919050505050808060010191505061071f565b610e97565b600090505b8251811015610c125782818151811015156109bf57fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209150600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a83610b6887600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5886040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610b1f57600080fd5b505af1158015610b33573d6000803e3d6000fd5b505050506040513d6020811015610b4957600080fd5b81019080805190602001909291905050506154e490919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015610bc957600080fd5b505af1158015610bdd573d6000803e3d6000fd5b505050506040513d6020811015610bf357600080fd5b81019080805190602001909291905050505080806001019150506109a8565b600090505b86869050811015610e96578686828181101515610c3057fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209150600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a83610dec87600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5886040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015610da357600080fd5b505af1158015610db7573d6000803e3d6000fd5b505050506040513d6020811015610dcd57600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015610e4d57600080fd5b505af1158015610e61573d6000803e3d6000fd5b505050506040513d6020811015610e7757600080fd5b8101908080519060200190929190505050508080600101915050610c17565b5b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015610fb157600080fd5b505af1158015610fc5573d6000803e3d6000fd5b505050506040513d6020811015610fdb57600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ec672cf63360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902089896040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250602002808284378201915050945050505050602060405180830381600087803b15801561112357600080fd5b505af1158015611137573d6000803e3d6000fd5b505050506040513d602081101561114d57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167fd46c1e81ba0ed24eccce499e94541a147c4fa3f6346854d76d70f3b6d401f8db8888604051808060200182810382528484828181526020019250602002808284378201915050935050505060405180910390a2600194505050505092915050565b600080600080600060606002600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156112f557600080fd5b505af1158015611309573d6000803e3d6000fd5b505050506040513d602081101561131f57600080fd5b810190808051906020019092919050505014151561133c57600080fd5b3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902094506001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5876040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561145257600080fd5b505af1158015611466573d6000803e3d6000fd5b505050506040513d602081101561147c57600080fd5b810190808051906020019092919050505014151561149957600080fd5b3360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209350600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5856040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156115ad57600080fd5b505af11580156115c1573d6000803e3d6000fd5b505050506040513d60208110156115d757600080fd5b810190808051906020019092919050505092503360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209150600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f733836040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b1580156116fe57600080fd5b505af1158015611712573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f82011682018060405250602081101561173c57600080fd5b81019080805164010000000081111561175457600080fd5b8281019050602081018481111561176a57600080fd5b815185602082028301116401000000008211171561178757600080fd5b5050929190505050905061179e81338560006154fd565b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635beff993836040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561183757600080fd5b505af115801561184b573d6000803e3d6000fd5b505050506040513d602081101561186157600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8660026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561191557600080fd5b505af1158015611929573d6000803e3d6000fd5b505050506040513d602081101561193f57600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf856040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156119ea57600080fd5b505af11580156119fe573d6000803e3d6000fd5b505050506040513d6020811015611a1457600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167f46f83f2216482f0d8839e93d2c2616a5fb914d68342883e58d8842b441bd54d4426040518082815260200191505060405180910390a260019550505050505090565b600080600080606060038060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611b9a57600080fd5b505af1158015611bae573d6000803e3d6000fd5b505050506040513d6020811015611bc457600080fd5b8101908080519060200190929190505050141515611be157600080fd5b3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902093506001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5866040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611cf757600080fd5b505af1158015611d0b573d6000803e3d6000fd5b505050506040513d6020811015611d2157600080fd5b8101908080519060200190929190505050141515611d3e57600080fd5b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611e4f57600080fd5b505af1158015611e63573d6000803e3d6000fd5b505050506040513d6020811015611e7957600080fd5b81019080805190602001909291905050509250600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166321f8a7213360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015611f9d57600080fd5b505af1158015611fb1573d6000803e3d6000fd5b505050506040513d6020811015611fc757600080fd5b81019080805190602001909291905050509150600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614151561247857600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902061222586600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58860405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156121dc57600080fd5b505af11580156121f0573d6000803e3d6000fd5b505050506040513d602081101561220657600080fd5b81019080805190602001909291905050506154e490919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561228657600080fd5b505af115801561229a573d6000803e3d6000fd5b505050506040513d60208110156122b057600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b1580156123d357600080fd5b505af11580156123e7573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f82011682018060405250602081101561241157600080fd5b81019080805164010000000081111561242957600080fd5b8281019050602081018481111561243f57600080fd5b815185602082028301116401000000008211171561245c57600080fd5b5050929190505050905061247381838560006154fd565b61262a565b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7333360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b15801561258957600080fd5b505af115801561259d573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f8201168201806040525060208110156125c757600080fd5b8101908080516401000000008111156125df57600080fd5b828101905060208101848111156125f557600080fd5b815185602082028301116401000000008211171561261257600080fd5b5050929190505050905061262981338560006154fd565b5b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8560026040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156126cc57600080fd5b505af11580156126e0573d6000803e3d6000fd5b505050506040513d60208110156126f657600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e756e766f746554696d650000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020426040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561282157600080fd5b505af1158015612835573d6000803e3d6000fd5b505050506040513d602081101561284b57600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635beff9933360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561296e57600080fd5b505af1158015612982573d6000803e3d6000fd5b505050506040513d602081101561299857600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16630e14a3763360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612abb57600080fd5b505af1158015612acf573d6000803e3d6000fd5b505050506040513d6020811015612ae557600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf3360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612c0857600080fd5b505af1158015612c1c573d6000803e3d6000fd5b505050506040513d6020811015612c3257600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff167f2199fcd2dd4d48edeef57d851b15e6ef618dd40ae1104d8e753602502e33ef1c426040518082815260200191505060405180910390a2600194505050505090565b600080600080606060003360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902094506000600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5876040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612dbd57600080fd5b505af1158015612dd1573d6000803e3d6000fd5b505050506040513d6020811015612de757600080fd5b8101908080519060200190929190505050141515612e0457600080fd5b6001543410151515612e1557600080fd5b600073ffffffffffffffffffffffffffffffffffffffff168973ffffffffffffffffffffffffffffffffffffffff16141515613285576002600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58b60405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015612f5e57600080fd5b505af1158015612f72573d6000803e3d6000fd5b505050506040513d6020811015612f8857600080fd5b8101908080519060200190929190505050141515612fa557600080fd5b6001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58b60405180807f70726f78792e7374617475730000000000000000000000000000000000000000815250600c018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156130b857600080fd5b505af11580156130cc573d6000803e3d6000fd5b505050506040513d60208110156130e257600080fd5b81019080805190602001909291905050501415156130ff57600080fd5b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ca446dd93360405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390208b6040518363ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018083600019166000191681526020018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200192505050602060405180830381600087803b15801561324457600080fd5b505af1158015613258573d6000803e3d6000fd5b505050506040513d602081101561326e57600080fd5b810190808051906020019092919050505050613430565b6132bd888880806020026020016040519081016040528093929190818152602001838360200280828437820191505050505050614e65565b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663ec672cf63360405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390208a8a6040518463ffffffff167c01000000000000000000000000000000000000000000000000000000000281526004018084600019166000191681526020018060200182810382528484828181526020019250602002808284378201915050945050505050602060405180830381600087803b1580156133f357600080fd5b505af1158015613407573d6000803e3d6000fd5b505050506040513d602081101561341d57600080fd5b8101908080519060200190929190505050505b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8660036040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156134d257600080fd5b505af11580156134e6573d6000803e3d6000fd5b505050506040513d60208110156134fc57600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e7374616b6564000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020346040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561362757600080fd5b505af115801561363b573d6000803e3d6000fd5b505050506040513d602081101561365157600080fd5b81019080805190602001909291905050505061366c34615870565b9350600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020866040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561378757600080fd5b505af115801561379b573d6000803e3d6000fd5b505050506040513d60208110156137b157600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902060016040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156138dd57600080fd5b505af11580156138f1573d6000803e3d6000fd5b505050506040513d602081101561390757600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a3360405180807f766f74652e756e766f746554696d650000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206000546040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613a3457600080fd5b505af1158015613a48573d6000803e3d6000fd5b505050506040513d6020811015613a5e57600080fd5b810190808051906020019092919050505050600073ffffffffffffffffffffffffffffffffffffffff168973ffffffffffffffffffffffffffffffffffffffff16141515613e99578860405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209250600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a84613c4687600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5896040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015613bfd57600080fd5b505af1158015613c11573d6000803e3d6000fd5b505050506040513d6020811015613c2757600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b158015613ca757600080fd5b505af1158015613cbb573d6000803e3d6000fd5b505050506040513d6020811015613cd157600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338a60405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b158015613df457600080fd5b505af1158015613e08573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f820116820180604052506020811015613e3257600080fd5b810190808051640100000000811115613e4a57600080fd5b82810190506020810184811115613e6057600080fd5b8151856020820283011164010000000082111715613e7d57600080fd5b50509291905050509150613e94828a8660016154fd565b61411e565b600090505b8787905081101561411d578787828181101515613eb757fe5b9050602002013573ffffffffffffffffffffffffffffffffffffffff1660405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390209250600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a8461407387600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5896040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561402a57600080fd5b505af115801561403e573d6000803e3d6000fd5b505050506040513d602081101561405457600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b6040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b1580156140d457600080fd5b505af11580156140e8573d6000803e3d6000fd5b505050506040513d60208110156140fe57600080fd5b8101908080519060200190929190505050508080600101915050613e9e565b5b8873ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167f5fd2438cc54f76807d5a8071c168927902854cef6c0c22176ded5bbe7174a9fb8a8a3460405180806020018381526020018281038252858582818152602001925060200280828437820191505094505050505060405180910390a36001955050505050509392505050565b60008060003360405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902092506003600102836000191614151561424e57600080fd5b3360405180807f766f74652e737461747573000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c01000000000000000000000000028152601401915050604051809103902091506002600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5846040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561436457600080fd5b505af1158015614378573d6000803e3d6000fd5b505050506040513d602081101561438e57600080fd5b81019080805190602001909291905050501415156143ab57600080fd5b61450b600254600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e756e766f746554696d650000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156144c257600080fd5b505af11580156144d6573d6000803e3d6000fd5b505050506040513d60208110156144ec57600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b4211151561451857600080fd5b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f53360405180807f766f74652e7374616b6564000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561462957600080fd5b505af115801561463d573d6000803e3d6000fd5b505050506040513d602081101561465357600080fd5b81019080805190602001909291905050509050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf846040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156146ff57600080fd5b505af1158015614713573d6000803e3d6000fd5b505050506040513d602081101561472957600080fd5b810190808051906020019092919050505050600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2b202bf836040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156147d457600080fd5b505af11580156147e8573d6000803e3d6000fd5b505050506040513d60208110156147fe57600080fd5b8101908080519060200190929190505050503373ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015614856573d6000803e3d6000fd5b503373ffffffffffffffffffffffffffffffffffffffff167fd2a1d21b9fe9c327ee10aa0097ad63335aabc0359e03d12808fc1abd1bf6ffa98242604051808381526020018281526020019250505060405180910390a2505050565b6000606060008060006060600080600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166321f8a7218a60405180807f766f74652e766f746550726f7879000000000000000000000000000000000000815250600e018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b1580156149d157600080fd5b505af11580156149e5573d6000803e3d6000fd5b505050506040513d60208110156149fb57600080fd5b81019080805190602001909291905050509350600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635948f7338a60405180807f766f74652e766f746550726f64756365727300000000000000000000000000008152506012018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050600060405180830381600087803b158015614b1f57600080fd5b505af1158015614b33573d6000803e3d6000fd5b505050506040513d6000823e3d601f19601f820116820180604052506020811015614b5d57600080fd5b810190808051640100000000811115614b7557600080fd5b82810190506020810184811115614b8b57600080fd5b8151856020820283011164010000000082111715614ba857600080fd5b50509291905050509250600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58a60405180807f766f74652e7374616b6564000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614cc357600080fd5b505af1158015614cd7573d6000803e3d6000fd5b505050506040513d6020811015614ced57600080fd5b81019080805190602001909291905050509150600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f58a60405180807f766f74652e776569676874000000000000000000000000000000000000000000815250600b018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614e1157600080fd5b505af1158015614e25573d6000803e3d6000fd5b505050506040513d6020811015614e3b57600080fd5b81019080805190602001909291905050509050838383839750975097509750505050509193509193565b6000601e825111151515614e7857600080fd5b6001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5846000815181101515614ec757fe5b9060200190602002015160405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b158015614fa357600080fd5b505af1158015614fb7573d6000803e3d6000fd5b505050506040513d6020811015614fcd57600080fd5b8101908080519060200190929190505050141515614fea57600080fd5b6001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f584600081518110151561503957fe5b9060200190602002015160405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561511557600080fd5b505af1158015615129573d6000803e3d6000fd5b505050506040513d602081101561513f57600080fd5b810190808051906020019092919050505014151561515c57600080fd5b600190505b81518110156154c4576001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f584848151811015156151b857fe5b9060200190602002015160405180807f757365722e726f6c6500000000000000000000000000000000000000000000008152506009018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561529457600080fd5b505af11580156152a8573d6000803e3d6000fd5b505050506040513d60208110156152be57600080fd5b81019080805190602001909291905050501415156152db57600080fd5b6001600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5848481518110151561532957fe5b9060200190602002015160405180807f70726f64756365722e7374617475730000000000000000000000000000000000815250600f018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c0100000000000000000000000002815260140191505060405180910390206040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561540557600080fd5b505af1158015615419573d6000803e3d6000fd5b505050506040513d602081101561542f57600080fd5b810190808051906020019092919050505014151561544c57600080fd5b818181518110151561545a57fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff16826001830381518110151561548b57fe5b9060200190602002015173ffffffffffffffffffffffffffffffffffffffff161015156154b757600080fd5b8080600101915050615161565b5050565b600081830190508281101515156154db57fe5b80905092915050565b60008282111515156154f257fe5b818303905092915050565b60008060008090505b865181101561586757868181518110151561551d57fe5b9060200190602002015160405180807f70726f64756365722e766f7465576569676874000000000000000000000000008152506013018273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166c010000000000000000000000000281526014019150506040518091039020915083156156945761568d85600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5856040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561564457600080fd5b505af1158015615658573d6000803e3d6000fd5b505050506040513d602081101561566e57600080fd5b81019080805190602001909291905050506154c890919063ffffffff16565b925061577d565b61577a85600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bd02d0f5856040518263ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401808260001916600019168152602001915050602060405180830381600087803b15801561573157600080fd5b505af1158015615745573d6000803e3d6000fd5b505050506040513d602081101561575b57600080fd5b81019080805190602001909291905050506154e490919063ffffffff16565b92505b600360009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e2a4853a83856040518363ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180836000191660001916815260200182815260200192505050602060405180830381600087803b15801561581e57600080fd5b505af1158015615832573d6000803e3d6000fd5b505050506040513d602081101561584857600080fd5b8101908080519060200190929190505050508080600101915050615506565b50505050505050565b600061588660018361588d90919063ffffffff16565b9050919050565b6000808314156158a057600090506158bf565b81830290508183828115156158b157fe5b041415156158bb57fe5b8090505b929150505600a165627a7a72305820fc9dac9cca8c5428ce7d6be3397a283dbcbc54ed32ec930f869e703c3797053b0029"
+    
+    // System contract abi generated by abigen
+    mainSystemContractABI = "[{\"constant\":false,\"inputs\":[],\"name\":\"getVoteSystemContract\",\"outputs\":[{\"name\":\"\",\"type\":\"address\"}],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"constant\":false,\"inputs\":[],\"name\":\"getRegSystemContract\",\"outputs\":[{\"name\":\"\",\"type\":\"address\"}],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"function\"}]"
+    regSystemContractABI  = "[{\"constant\":true,\"inputs\":[],\"name\":\"getAllProducersInfo\",\"outputs\":[{\"name\":\"\",\"type\":\"address[]\"},{\"name\":\"\",\"type\":\"uint256[]\"},{\"name\":\"\",\"type\":\"uint256\"}],\"payable\":false,\"stateMutability\":\"view\",\"type\":\"function\"}]"
+    voteSystemContractABI = ""
+    
+    // Method ID
+    getRegSystemContractMethodID  = "0xa79413e1"
+    getVoteSystemContractMethodID = "0x2083c5f9"
+    getAllProducersInfoMethodID   = "0x1af55073"
+)
+
+// System contract variables
+var (
+    mainSystemContractAddr = common.BytesToAddress([]byte{9})
+    regSystemContractAddr  = common.BytesToAddress([]byte{10})
+    voteSystemContractAddr = common.BytesToAddress([]byte{11})
+    
+    // Main contract variables
+    regContractBoolStorageIndex = common.HexToHash("0xddefb938278018c16c1e429eff219977102bbd1a16c0a9ca12f10ae5d425e820")
+    regContractBoolStorageValue = common.BytesToHash([]byte{1})
+    
+    regContractAddressStorageIndex = common.HexToHash("0x11cde89d6b0e279583592f4b5d7e2a720237412a0a98c492b67e3246c187c00f")
+    regContractAddressStorageValue = common.BytesToHash([]byte{10})
+    
+    voteContractBoolStorageIndex = common.HexToHash("0xc6eca4648b4c9f4004f098d97590d949321f0016aa3c54114fe524197e35c089")
+    voteContractBoolStorageValue = common.BytesToHash([]byte{1})
+    
+    voteContractAddressStorageIndex = common.HexToHash("0x31e02a136b0147c5e92de4e9eaf91014b31370d815bccd7fef5f673bbae5f4ad")
+    voteContractAddressStorageValue = common.BytesToHash([]byte{11})
+    
+    
+    // Reg contract variables
+    depositForJoinStorageIndex = common.BytesToHash([]byte{0})
+    depositForJoinStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000de0b6b3a7640000")
+    
+    lockTimeForDepositStorageIndex = common.BytesToHash([]byte{1})
+    lockTimeForDepositStorageValue = common.HexToHash("0x000000000000000000000000000000000000000000000000000000000003f480")
+    
+    lengthOFEpochStorageIndex = common.BytesToHash([]byte{2})
+    lengthOFEpochStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000004")
+    
+    initOutTimeStorageIndex = common.BytesToHash([]byte{3})
+    initOutTimeStorageValue = common.HexToHash("0x00000000000000000000000000000000000000000000000000000007915e5b80")
+    
+    systemContractStorageIndex = common.BytesToHash([]byte{4})
+    systemContractStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009")
+    
+    producerOpStorageIndex = common.BytesToHash([]byte{5})
+    producerOpStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009")
+    
+    
+    // Vote contract variables
+    outTimeStorageIndex = common.BytesToHash([]byte{0})
+    outTimeStorageValue = common.HexToHash("0x00000000000000000000000000000000000000000000000000000007915e5b80")
+    
+    leastDepositForVoteStorageIndex = common.BytesToHash([]byte{1})
+    leastDepositForVoteStorageValue = common.HexToHash("0x000000000000000000000000000000000000000000000000000000000003f480")
+    
+    lockTimeForVoteStorageIndex = common.BytesToHash([]byte{2})
+    lockTimeForVoteStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000de0b6b3a7640000")
+    
+    mainContractStorageIndex = common.BytesToHash([]byte{3})
+    mainContractStorageValue = common.HexToHash("0x0000000000000000000000000000000000000000000000000000000000000009")
+)
+
+// Upgradeable system contract for dpos
+// Producer/Voters can send tx to system contract for reg, unreg, vote and so on.
+var HardcodedContractsDpos = []HardcodedContract{
+    NewSystemContract(
+        mainSystemContractAddr,
+        mainSystemContractCode,
+        mainSystemContractABI,
+        // Set reg and vote system contract address
+        map[common.Hash]common.Hash{
+            // Set reg system contract bool; mapping key is : 0x3742e890ccd7b17f07ea3d26ddabc6a24bd77cb109451e8c225eced73747e219
+            regContractBoolStorageIndex: regContractBoolStorageValue,
+            // Set vote system contract bool; mapping key is : 0x46a34feee9b84783782542ca731df2e0a2b7305bf7c1dd556e25d64fd9fa0bd2
+            voteContractBoolStorageIndex: voteContractBoolStorageValue,
+            // Set reg system contract address
+            regContractAddressStorageIndex: voteContractBoolStorageIndex,
+            // Set vote system contract address
+            voteContractAddressStorageIndex: voteContractAddressStorageValue,
+        },
+    ),
+    NewSystemContract(
+        regSystemContractAddr,
+        regSystemContractCode,
+        regSystemContractABI,
+        map[common.Hash]common.Hash{
+            // Set init variables
+            depositForJoinStorageIndex: depositForJoinStorageValue,
+            lockTimeForDepositStorageIndex: lockTimeForDepositStorageValue,
+            lengthOFEpochStorageIndex: lengthOFEpochStorageValue,
+            initOutTimeStorageIndex: initOutTimeStorageValue,
+            systemContractStorageIndex: systemContractStorageValue,
+            producerOpStorageIndex: producerOpStorageValue,
+        },
+    ),
+    NewSystemContract(
+        voteSystemContractAddr,
+        voteSystemContractCode,
+        voteSystemContractABI,
+        map[common.Hash]common.Hash{
+            // Set init variables
+            outTimeStorageIndex: outTimeStorageValue,
+            leastDepositForVoteStorageIndex: leastDepositForVoteStorageValue,
+            lockTimeForVoteStorageIndex: lockTimeForVoteStorageValue,
+            mainContractStorageIndex: mainContractStorageValue,
+        },
+    ),
+}
+
+type HardcodedContract interface {
+    GetCode() string
+    GetContractAddr() common.Address
+    GetStorage() map[common.Hash]common.Hash
+}
+
+func NewSystemContract(addr common.Address, code string, abiStr string, storageSet map[common.Hash]common.Hash) *SystemContract {
+    // Ignore error
+    parsed, _ := abi.JSON(strings.NewReader(abiStr))
+    return &SystemContract{
+        Addr:       addr,
+        Code:       code,
+        StorageSet: storageSet,
+        ABI:        parsed,
+    }
+}
+
+// Main dpos upgradeable system contract entry.
+// Producer can send tx to this contract for proposing upgrade system
+// contract, the storage data will leave on this contract.
+type SystemContract struct {
+    Addr       common.Address
+    Code       string
+    StorageSet map[common.Hash]common.Hash
+    ABI        abi.ABI
+}
+
+func (m *SystemContract) GetCode() string {
+    return m.Code
+}
+
+func (m *SystemContract) GetContractAddr() common.Address {
+    return m.Addr
+}
+
+func (m *SystemContract) GetStorage() map[common.Hash]common.Hash {
+    return m.StorageSet
+}
+
+func (m *SystemContract) GetRegSystemContract(data []byte) common.Address {
+    var res = new(common.Address)
+    m.ABI.Unpack(res, "getCurrentSystemContractAddress", data)
+    return *res
+}
+
+// Caller for system contract which can get current reg system contract and get all producers.
+type SystemContractCaller struct {
+    mainSystemContractABI abi.ABI
+    regSystemContractABI  abi.ABI
+}
+
+func NewSystemContractCaller() *SystemContractCaller {
+    // Ignore error
+    mainABI, _ := abi.JSON(strings.NewReader(mainSystemContractABI))
+    regABI, _ := abi.JSON(strings.NewReader(regSystemContractABI))
+    return &SystemContractCaller{
+        mainABI,
+        regABI,
+    }
+}
+
+func (s *SystemContractCaller) GetRegSystemContractCall(header *types.Header) SystemCall {
+    copyAddr := mainSystemContractAddr
+    return NewCallMsg(&copyAddr, common.Hex2Bytes(getRegSystemContractMethodID), header.Number.Uint64())
+}
+
+func (s *SystemContractCaller) GetRegSystemContractAddress(data []byte) common.Address {
+    var res = new(common.Address)
+    s.mainSystemContractABI.Unpack(res, "getRegSystemContract", data)
+    return *res
+}
+
+func (s *SystemContractCaller) GetAllProducersInfoCall(header *types.Header, contract *common.Address) SystemCall {
+    return NewCallMsg(contract, common.Hex2Bytes(getAllProducersInfoMethodID), header.Number.Uint64())
+}
+
+func (s *SystemContractCaller) GetAllProducersInfo(data []byte) ([]common.Address, []*big.Int, *big.Int, error) {
+    var (
+        ret0 = new([]common.Address)
+        ret1 = new([]*big.Int)
+        ret2 = new(*big.Int)
+    )
+    out := &[]interface{}{
+        ret0,
+        ret1,
+        ret2,
+    }
+    
+    err := s.regSystemContractABI.Unpack(out, "getAllProducersInfo", data)
+    return *ret0, *ret1, *ret2, err
+}
+
+func NewCallMsg(addr *common.Address, data []byte, blockNum uint64) SystemCall {
+    return SystemCall{
+        ToAddr:    addr,
+        DataField: data,
+        GasSupply: 50000000,
+        AtBlock:   blockNum,
+    }
+}
+
+type SystemCall struct {
+    ToAddr    *common.Address
+    DataField []byte
+    GasSupply uint64
+    AtBlock   uint64
+}
+
+func (s SystemCall) From() common.Address { return common.Address{} }
+func (s SystemCall) Nonce() uint64        { return 0 }
+func (s SystemCall) CheckNonce() bool     { return false }
+func (s SystemCall) To() *common.Address  { return s.ToAddr }
+func (s SystemCall) GasPrice() *big.Int   { return new(big.Int).SetUint64(0) }
+func (s SystemCall) Gas() uint64          { return s.GasSupply }
+func (s SystemCall) Value() *big.Int      { return new(big.Int).SetUint64(0) }
+func (s SystemCall) Data() []byte         { return s.DataField }
+
